// src/http-courses/models/modelsEmprego.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * API Go
 * API de serviços para aplicativos da Prefeitura do Rio
 * OpenAPI spec version: 1.0
 */
import type { ModelsEmpresa } from './modelsEmpresa'
import type { ModelsEscolaridade } from './modelsEscolaridade'
import type { ModelsJornadaTrabalho } from './modelsJornadaTrabalho'
import type { ModelsStatusEmprego } from './modelsStatusEmprego'
import type { ModelsTipoContratacao } from './modelsTipoContratacao'
import type { ModelsTurno } from './modelsTurno'

export interface ModelsEmprego {
  beneficios?: string
  contato_duvidas?: string
  created_at?: string
  data_inicio_prevista?: string
  data_limite_candidatura?: string
  descricao?: string
  /** Relacionamentos */
  empresa?: ModelsEmpresa
  empresa_id?: number
  escolaridade?: ModelsEscolaridade
  escolaridade_id?: number
  id?: number
  jornada_trabalho?: ModelsJornadaTrabalho
  latitude?: number
  longitude?: number
  numero_vagas?: number
  orgao_id?: string
  pre_requisitos?: string
  salario_max?: number
  salario_min?: number
  status?: ModelsStatusEmprego
  tipo_contratacao?: ModelsTipoContratacao
  titulo?: string
  turno?: ModelsTurno
  updated_at?: string
}